More output patches

+++ src/MP3Track.cpp
@@ -37,7 +37,7 @@ void MP3Track::load()
 void MP3Track::analyze_beatgrid()
 {
     std::cout << "[MP3Track::analyze_beatgrid] Analyzing beat grid for: \"" << title << "\"\n";
-    double beats = (duration_seconds / 60.0) * bpm;
+    int beats = (duration_seconds / 60.0) * bpm;
     double precision_factor = bitrate / 320.0;
 
     std::cout << "  → Estimated beats: " << beats
